// app/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  SUPERADMIN: 'SUPERADMIN',
  ADMIN: 'ADMIN',
  MANAGER: 'MANAGER',
  ATTENDANCE_INCHARGE: 'ATTENDANCE_INCHARGE',
  JANAB: 'JANAB',
  MUMIN: 'MUMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const Gender = {
  MALE: 'MALE',
  FEMALE: 'FEMALE'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]


export const EnrollmentStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type EnrollmentStatus = (typeof EnrollmentStatus)[keyof typeof EnrollmentStatus]


export const AttendanceMethod = {
  QR_SCAN: 'QR_SCAN',
  MANUAL_ENTRY: 'MANUAL_ENTRY',
  LOCATION_BASED_SELF: 'LOCATION_BASED_SELF'
} as const

export type AttendanceMethod = (typeof AttendanceMethod)[keyof typeof AttendanceMethod]


export const EmailStatus = {
  PENDING: 'PENDING',
  SENT: 'SENT',
  FAILED: 'FAILED'
} as const

export type EmailStatus = (typeof EmailStatus)[keyof typeof EmailStatus]


export const RequestStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type RequestStatus = (typeof RequestStatus)[keyof typeof RequestStatus]